[Refactor]: private model store
